added styles and responsiveness to sidebar
