+++ src/App.css
@@ -4,7 +4,7 @@ label {
 }
 
 main {
-  margin-left: 170px;
+  margin-left: 190px;
 }
 
 ul {
