+++ src/styles/Sidebar.module.css
@@ -1,0 +1,62 @@
+.hamburger {
+  position: fixed;
+  z-index: 2;
+  top: 20px;
+  left: 20px;
+  outline: none;
+  border: none;
+  background-color: transparent;
+}
+
+.sidenav {
+  height: 100%;
+  width: 190px;
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  left: 0;
+  padding-left: 10px;
+  overflow-x: hidden;
+  padding-top: 50px;
+  border-right: 1px solid lightgray;
+  background-color: white;
+}
+
+.toggle {
+  display: none;
+}
+
+.navlink {
+  padding: 6px 8px 6px 16px;
+  text-decoration: none;
+  font-size: 25px;
+  color: black;
+  display: block;
+}
+
+.active {
+  color: white;
+  background-color: rgba(150, 191, 1);
+}
+
+.navlink:hover {
+  background-color: rgba(150, 191, 1, 0.7);
+  color: white;
+  text-decoration: none;
+}
+
+.navlink:active {
+  color: white;
+  background-color: rgba(150, 191, 1);
+}
+
+.main {
+  margin-left: 160px; /* Same as the width of the sidenav */
+  font-size: 28px; /* Increased text to enable scrolling */
+  padding: 0 10px;
+}
+
+@media (min-width: 600px) {
+  .sidenav { display: block; }
+  .hamburger { display: none; }
+}
